Move dynamic compression to shared code
