+++ src/Servers/Kestrel/shared/test/PipeWriterHttp2FrameExtensions.cs
@@ -8,7 +8,6 @@ using System.Collections.Generic;
 using System.IO.Pipelines;
 using System.Net.Http.HPack;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2.HPack;
 
 namespace Microsoft.AspNetCore.Testing
 {
@@ -26,13 +25,13 @@ namespace Microsoft.AspNetCore.Testing
             writer.Write(payload);
         }
 
-        public static void WriteStartStream(this PipeWriter writer, int streamId, Http2HPackEncoder hpackEncoder, Http2HeadersEnumerator headers, byte[] headerEncodingBuffer, bool endStream, Http2Frame frame = null)
+        public static void WriteStartStream(this PipeWriter writer, int streamId, HPackEncoder hpackEncoder, Http2HeadersEnumerator headers, byte[] headerEncodingBuffer, bool endStream, Http2Frame frame = null)
         {
             frame ??= new Http2Frame();
             frame.PrepareHeaders(Http2HeadersFrameFlags.NONE, streamId);
 
             var buffer = headerEncodingBuffer.AsSpan();
-            var done = hpackEncoder.BeginEncodeHeaders(headers, buffer, out var length);
+            var done = HPackHeaderWriter.BeginEncodeHeaders(hpackEncoder, headers, buffer, out var length);
             frame.PayloadLength = length;
 
             if (done)
@@ -52,7 +51,7 @@ namespace Microsoft.AspNetCore.Testing
             {
                 frame.PrepareContinuation(Http2ContinuationFrameFlags.NONE, streamId);
 
-                done = hpackEncoder.ContinueEncodeHeaders(headers, buffer, out length);
+                done = HPackHeaderWriter.ContinueEncodeHeaders(hpackEncoder, headers, buffer, out length);
                 frame.PayloadLength = length;
 
                 if (done)
